Update War3Net to v6.0.3 (#98)
- Added `Count` and `EnumerateHashes()` to `MpqArchive`.
- Changed `MpqArchive[int]` from internal to public.

+++ src/War3Net.IO.Mpq/MpqTable.cs
@@ -20,7 +20,7 @@
         /// <summary>
         /// Gets the capacity of the <see cref="MpqTable"/>.
         /// </summary>
-        public abstract uint Size { get; }
+        public abstract int Size { get; }
 
         /// <summary>
         /// Gets the key used to encrypt and decrypt the <see cref="MpqTable"/>.
@@ -106,7 +106,7 @@
 
                 using (var reader = new BinaryReader(memoryStream))
                 {
-                    data = reader.ReadBytes((int)Size * EntrySize);
+                    data = reader.ReadBytes(Size * EntrySize);
                     Encrypt(data);
                 }
             }
